extracted filter query builder into a function
