+++ src/cli.js
@@ -12,7 +12,7 @@ const updateNotifier = require('update-notifier')
 const pkg = require('../package.json')
 
 const tasks = require('./tasks')
-const { getQuestions, lastStep, api, creds } = require('./utils')
+const { getQuestions, lastStep, api, creds, buildFilterQuery } = require('./utils')
 const { SYNC_TYPES, COMMANDS } = require('./constants')
 
 clear()
@@ -292,27 +292,7 @@ program
         }
       })
 
-      let filterQuery
-      if (filter) {
-        const operators = ['is', 'in', 'not_in', 'like', 'not_like', 'any_in_array', 'all_in_array', 'gt_date', 'lt_date', 'gt_int', 'lt_int', 'gt_float', 'lt_float']
-        if (!keys || !operations || !values) {
-          throw new Error('Filter options are required: --keys; --operations; --values')
-        }
-        const _keys = keys.split(' ')
-        const _operations = operations.split(' ')
-        const _values = values.split(' ')
-        if (_keys.length !== _operations.length || _keys.length !== _values.length) {
-          throw new Error('The number of keys, operations and values must be the same')
-        }
-        const invalidOperators = _operations.filter((o) => !operators.includes(o))
-        if (invalidOperators.length) {
-          throw new Error('Invalid operator(s) applied for filter: ' + invalidOperators.join(' '))
-        }
-        filterQuery = {}
-        _keys.forEach((key, index) => {
-          filterQuery[key] = { [_operations[index]]: _values[index] }
-        })
-      }
+      const filterQuery = filter ? buildFilterQuery(keys, operations, values) : undefined
 
       const token = creds.get().token || null
       await tasks.sync(_types, {
